requirements.txt
openai>=0.27.4
python-dotenv>=0.19.2
Flask>=2.0.2
black>=23.3.0